Missed cast

+++ src/main/java/us/q3q/fido2/FIDO2Applet.java
@@ -995,7 +995,7 @@ public final class FIDO2Applet extends Applet implements ExtendedLength {
                 } else if (targetRKSlot < numResidentCredentials) {
                     // We need to put the credential at the end of the list so it's still the "most recent" one
                     for (short i = targetRKSlot; i < (short)(numResidentCredentials - 1); i++) {
-                        residentKeys[i] = residentKeys[i+1];
+                        residentKeys[i] = residentKeys[(short)(i + 1)];
                     }
                     targetRKSlot = (short)(numResidentCredentials - 1);
                 }
